src/requirements.txt
paho-mqtt==1.5.0
python-telegram-bot==12.8
